backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
pydantic
python-multipart
passlib[bcrypt]
python-jose
